+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.2
+
+* Renamed `fromProtocol` to `fromApiObject` in `Key` and `Entity` to follow
+  common naming convention
+* `fromValue` now handles empty values (treats them as `null`)
+* `Entity` throws more meaningful errors in `fromApiObject`
+
 ## 0.1.1+3
 
 * Added `createKey` method to `EntityBridge`